requirements.txt
streamlit>=1.28.0
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
chromadb>=0.4.15
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.17.0
streamlit-authenticator>=0.2.3
lxml>=4.9.0

